Initial commit
Updated README and gitignore

First Project: Vowel Recognition

The aim of this project is to create a little program to find which
vowel was said in a small sample of audio. This is done by recording
a little sample of audio and throug spectral analysis (with FFT)
determine the two formants. With this two formants we can determine
which vowel was recorded.